+++ jommerce/auth/models.py
@@ -1,5 +1,7 @@
+from django.utils.translation import gettext_lazy as _
 from django.contrib.auth.models import AbstractUser
 from django.contrib.auth.models import UserManager
+from django.db import models
 
 
 class CustomUserManager(UserManager):
@@ -7,6 +9,8 @@ class CustomUserManager(UserManager):
 
 
 class CustomUser(AbstractUser):
+    email = models.EmailField(_("email address"), unique=True, null=True)
+
     objects = CustomUserManager()
 
     class Meta(AbstractUser.Meta):