+++ jommerce/project_template/project_name/settings/base.py
@@ -1,2 +1,0 @@
-ROOT_URLCONF = "{{ project_name }}.urls"
-WSGI_APPLICATION = "{{ project_name }}.wsgi.application"