+++ jommerce/urls.py
@@ -1,7 +1,7 @@
 """jommerce URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/3.2/topics/http/urls/
+    https://docs.djangoproject.com/en/4.0/topics/http/urls/
 Examples:
 Function views
     1. Add an import:  from my_app import views
@@ -15,9 +15,9 @@ Including another URLconf
 """
 from django.views.generic import TemplateView
 from django.conf.urls.static import static
+from django.urls import path, include
 from django.conf import settings
 from django.contrib import admin
-from django.urls import path
 
 urlpatterns = [
     path(
@@ -33,3 +33,11 @@ if settings.DEBUG:
     urlpatterns += static(
         settings.MEDIA_URL, document_root=settings.MEDIA_ROOT
     )
+    try:
+        import debug_toolbar
+    except ImportError:
+        pass
+    else:
+        urlpatterns += [
+            path("__debug__/", include(debug_toolbar.urls)),
+        ]