+++ CHANGELOG.md
@@ -6,16 +6,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.1.1] - 2021-11-12
+## [1.0.0] - 2021-12-14
+### Changed
+- Upgrade ```Django``` to 4.0
+- Use **environment variables** to read sensitive data instead of using the ```.secrets``` file.
+- The ```email``` field is ```unique``` by default.
+- The **project template** has changed.
+### Added
+- Add ```django-debug-toolbar``` to the development environment.
+- Add **PostgreSQL Database Settings** to the production environment.
+### Removed
+- The ```.secrets``` file has been deleted.
+
+## [0.1.1] - 2021-12-11
 ### Fixed
 - Fixed generate ```SECRET_KEY```
 - Fixed 'Page not found' error on ```Home Page```
 
-## [0.1.0] - 2021-08-12
+## [0.1.0] - 2021-12-08
 ### Added
 - Added ```Custom User Model```
 - Added ```auth``` app.
 - Added ```jommerce``` command.
 
+[1.0.0]: https://github.com/jommerce/jommerce/releases/tag/v1.0.0
 [0.1.1]: https://github.com/jommerce/jommerce/releases/tag/v0.1.1
 [0.1.0]: https://github.com/jommerce/jommerce/releases/tag/v0.1.0